deduplicate FixedExpressionValue by moving it to common-api
