+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/additionalconfig/AdditionalUIConfigProvider.scala
@@ -2,7 +2,7 @@ package pl.touk.nussknacker.ui.additionalconfig
 
 import pl.touk.nussknacker.engine.api.component._
 import pl.touk.nussknacker.engine.compile.nodecompilation.ValueEditorValidator
-import pl.touk.nussknacker.engine.graph.node.FragmentInputDefinition.FixedExpressionValue
+import pl.touk.nussknacker.engine.graph.expression.FixedExpressionValue
 import pl.touk.nussknacker.engine.graph.node.{ParameterValueCompileTimeValidation, ValueInputWithFixedValues}
 
 /**
