+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/definition/scenarioproperty/UiScenarioPropertyEditorDeterminerTest.scala
@@ -2,12 +2,9 @@ package pl.touk.nussknacker.ui.definition.scenarioproperty
 
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
-import pl.touk.nussknacker.engine.api.definition.{
-  FixedExpressionValue,
-  FixedValuesParameterEditor,
-  StringParameterEditor
-}
+import pl.touk.nussknacker.engine.api.definition.{FixedValuesParameterEditor, StringParameterEditor}
 import pl.touk.nussknacker.engine.api.component.ScenarioPropertyConfig
+import pl.touk.nussknacker.engine.graph.expression.FixedExpressionValue
 
 class UiScenarioPropertyEditorDeterminerTest extends AnyFunSuite with Matchers {
