+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/validation/UIProcessValidatorSpec.scala
@@ -22,7 +22,7 @@ import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.canonicalgraph.canonicalnode.{FlatNode, SplitNode}
 import pl.touk.nussknacker.engine.compile.ProcessValidator
 import pl.touk.nussknacker.engine.graph.EdgeType.{NextSwitch, SwitchDefault}
-import pl.touk.nussknacker.engine.graph.expression.Expression
+import pl.touk.nussknacker.engine.graph.expression.{Expression, FixedExpressionValue}
 import pl.touk.nussknacker.engine.graph.fragment.FragmentRef
 import pl.touk.nussknacker.engine.graph.node.FragmentInputDefinition.{FragmentClazzRef, FragmentParameter}
 import pl.touk.nussknacker.engine.graph.node._
@@ -473,11 +473,11 @@ class UIProcessValidatorSpec extends AnyFunSuite with Matchers {
                   "subParam1",
                   FragmentClazzRef[String],
                   required = false,
-                  initialValue = Some(FragmentInputDefinition.FixedExpressionValue("'outsidePreset'", "outsidePreset")),
+                  initialValue = Some(FixedExpressionValue("'outsidePreset'", "outsidePreset")),
                   hintText = None,
                   valueEditor = Some(
                     ValueInputWithFixedValuesProvided(
-                      fixedValuesList = List(FragmentInputDefinition.FixedExpressionValue("'someValue'", "someValue")),
+                      fixedValuesList = List(FixedExpressionValue("'someValue'", "someValue")),
                       allowOtherValue = false
                     )
                   ),
@@ -491,7 +491,7 @@ class UIProcessValidatorSpec extends AnyFunSuite with Matchers {
                   hintText = None,
                   valueEditor = Some(
                     ValueInputWithFixedValuesProvided(
-                      fixedValuesList = List(FragmentInputDefinition.FixedExpressionValue("'someValue'", "someValue")),
+                      fixedValuesList = List(FixedExpressionValue("'someValue'", "someValue")),
                       allowOtherValue = false
                     )
                   ),
@@ -623,7 +623,7 @@ class UIProcessValidatorSpec extends AnyFunSuite with Matchers {
                 hintText = None,
                 valueEditor = Some(
                   ValueInputWithFixedValuesProvided(
-                    fixedValuesList = List(FragmentInputDefinition.FixedExpressionValue("'someValue'", "someValue")),
+                    fixedValuesList = List(FixedExpressionValue("'someValue'", "someValue")),
                     allowOtherValue = false
                   )
                 ),
