+++ engine/flink/components/base/src/main/java/pl/touk/nussknacker/engine/flink/util/transformer/aggregate/AggregateHelper.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import pl.touk.nussknacker.engine.api.Hidden;
 import pl.touk.nussknacker.engine.api.ParamName;
 import pl.touk.nussknacker.engine.api.definition.DualParameterEditor;
-import pl.touk.nussknacker.engine.api.definition.FixedExpressionValue;
+import pl.touk.nussknacker.engine.graph.expression.FixedExpressionValue;
 import pl.touk.nussknacker.engine.api.definition.FixedValuesParameterEditor;
 import pl.touk.nussknacker.engine.api.definition.SimpleParameterEditor;
 import pl.touk.nussknacker.engine.api.editor.DualEditorMode;
