+++ engine/flink/management/dev-model/src/main/scala/pl/touk/nussknacker/engine/management/sample/DevProcessConfigCreator.scala
@@ -12,7 +12,6 @@ import org.apache.flink.streaming.api.functions.sink.DiscardingSink
 import pl.touk.nussknacker.engine.api._
 import pl.touk.nussknacker.engine.api.component.{ComponentGroupName, ParameterConfig, SingleComponentConfig}
 import pl.touk.nussknacker.engine.api.definition.{
-  FixedExpressionValue,
   FixedValuesParameterEditor,
   MandatoryParameterValidator,
   StringParameterEditor
@@ -24,6 +23,7 @@ import pl.touk.nussknacker.engine.flink.util.source.{
   ReturningClassInstanceSource,
   ReturningTestCaseClass
 }
+import pl.touk.nussknacker.engine.graph.expression.FixedExpressionValue
 import pl.touk.nussknacker.engine.kafka.consumerrecord.{
   ConsumerRecordToJsonFormatterFactory,
   FixedValueDeserializationSchemaFactory
