+++ engine/flink/management/dev-model/src/main/scala/pl/touk/nussknacker/engine/management/sample/service/DynamicMultipleParamsService.scala
@@ -7,16 +7,12 @@ import pl.touk.nussknacker.engine.api.context.transformation.{
   NodeDependencyValue,
   SingleInputGenericNodeTransformation
 }
-import pl.touk.nussknacker.engine.api.definition.{
-  FixedExpressionValue,
-  FixedValuesParameterEditor,
-  NodeDependency,
-  Parameter
-}
+import pl.touk.nussknacker.engine.api.definition.{FixedValuesParameterEditor, NodeDependency, Parameter}
 import pl.touk.nussknacker.engine.api.process.ComponentUseCase
 import pl.touk.nussknacker.engine.api.test.InvocationCollectors
 import pl.touk.nussknacker.engine.api.typed.typing.Typed
-import pl.touk.nussknacker.engine.graph.expression.Expression
+import pl.touk.nussknacker.engine.graph.expression.FixedExpressionValue.nullFixedValue
+import pl.touk.nussknacker.engine.graph.expression.{Expression, FixedExpressionValue}
 
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -32,7 +28,7 @@ object DynamicMultipleParamsService extends EagerService with SingleInputGeneric
         Some(
           FixedValuesParameterEditor(
             List(
-              FixedExpressionValue("", ""),
+              nullFixedValue,
               FixedExpressionValue("'fooValueFromConfig'", "From Config"),
               FixedExpressionValue("'other'", "Other")
             )
