+++ engine/flink/management/dev-model/src/main/scala/pl/touk/nussknacker/engine/management/sample/transformer/DynamicParametersMixin.scala
@@ -8,7 +8,6 @@ import pl.touk.nussknacker.engine.api.context.transformation.{
   SingleInputGenericNodeTransformation
 }
 import pl.touk.nussknacker.engine.api.definition.{
-  FixedExpressionValue,
   FixedValuesParameterEditor,
   NodeDependency,
   Parameter,
@@ -18,6 +17,7 @@ import pl.touk.nussknacker.engine.api.MetaData
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.CustomNodeError
 import pl.touk.nussknacker.engine.api.context.ValidationContext
 import pl.touk.nussknacker.engine.api.NodeId
+import pl.touk.nussknacker.engine.graph.expression.FixedExpressionValue
 
 trait DynamicParametersMixin extends SingleInputGenericNodeTransformation[AnyRef] {
