+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/compile/nodecompilation/ValueEditorValidator.scala
@@ -5,30 +5,24 @@ import cats.data.ValidatedNel
 import cats.implicits.toTraverseOps
 import pl.touk.nussknacker.engine.api.context.PartSubGraphCompilationError
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError._
-import pl.touk.nussknacker.engine.api.definition.{
-  DualParameterEditor,
-  FixedExpressionValue,
-  FixedValuesParameterEditor,
-  ParameterEditor
-}
+import pl.touk.nussknacker.engine.api.definition.{DualParameterEditor, FixedValuesParameterEditor, ParameterEditor}
 import pl.touk.nussknacker.engine.api.editor.DualEditorMode
-import pl.touk.nussknacker.engine.graph.node.FragmentInputDefinition.{
-  FixedExpressionValue => FragmentFixedExpressionValue
-}
+import pl.touk.nussknacker.engine.graph.expression.FixedExpressionValue
+import pl.touk.nussknacker.engine.graph.expression.FixedExpressionValue.nullFixedValue
 import pl.touk.nussknacker.engine.graph.node.ValueInputWithFixedValues
 
 object ValueEditorValidator {
 
   def validateAndGetEditor( // this method doesn't validate the compilation validity of FixedExpressionValues (it requires validationContext and expressionCompiler, see FragmentParameterValidator.validateFixedExpressionValues)
       valueEditor: ValueInputWithFixedValues,
-      initialValue: Option[FragmentFixedExpressionValue],
+      initialValue: Option[FixedExpressionValue],
       paramName: String,
       nodeIds: Set[String]
   ): ValidatedNel[PartSubGraphCompilationError, ParameterEditor] = {
     validateFixedValuesList(valueEditor, initialValue, paramName, nodeIds)
       .andThen { _ =>
         val fixedValuesEditor = FixedValuesParameterEditor(
-          nullFixedValue +: valueEditor.fixedValuesList.map(v => FixedExpressionValue(v.expression, v.label))
+          nullFixedValue +: valueEditor.fixedValuesList
         )
 
         if (valueEditor.allowOtherValue) {
@@ -39,11 +33,9 @@ object ValueEditorValidator {
       }
   }
 
-  private val nullFixedValue: FixedExpressionValue = FixedExpressionValue("", "")
-
   private def validateFixedValuesList(
       valueEditor: ValueInputWithFixedValues,
-      initialValue: Option[FragmentFixedExpressionValue],
+      initialValue: Option[FixedExpressionValue],
       paramName: String,
       nodeIds: Set[String]
   ): ValidatedNel[PartSubGraphCompilationError, Unit] =
@@ -60,7 +52,7 @@ object ValueEditorValidator {
 
   private def initialValueNotPresentInPossibleValues(
       valueEditor: ValueInputWithFixedValues,
-      initialValue: Option[FragmentFixedExpressionValue]
+      initialValue: Option[FixedExpressionValue]
   ) = initialValue.exists(!valueEditor.fixedValuesList.contains(_))
 
 }
