+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/definition/component/parameter/editor/EditorExtractor.scala
@@ -4,6 +4,8 @@ import pl.touk.nussknacker.engine.api.definition._
 import pl.touk.nussknacker.engine.api.editor.{DualEditor, DualEditorMode, RawEditor, SimpleEditor, SimpleEditorType}
 import pl.touk.nussknacker.engine.api.component.ParameterConfig
 import pl.touk.nussknacker.engine.definition.component.parameter.ParameterData
+import pl.touk.nussknacker.engine.graph.expression.FixedExpressionValue
+import pl.touk.nussknacker.engine.graph.expression.FixedExpressionValue.nullFixedValue
 import pl.touk.nussknacker.engine.graph.node.{ValueInputWithFixedValues, ValueInputWithFixedValuesProvided}
 
 object EditorExtractor {
@@ -11,7 +13,7 @@ object EditorExtractor {
   def extract(valueInput: ValueInputWithFixedValues): ParameterEditor = valueInput match {
     case ValueInputWithFixedValuesProvided(fixedValuesList, allowOtherValue) =>
       val fixedValuesEditor = FixedValuesParameterEditor(
-        nullFixedValue +: fixedValuesList.map(v => FixedExpressionValue(v.expression, v.label))
+        nullFixedValue +: fixedValuesList
       )
 
       if (allowOtherValue) {
@@ -26,8 +28,6 @@ object EditorExtractor {
       .orElse(extractFromAnnotations(param))
       .orElse(new ParameterTypeEditorDeterminer(param.typing).determine())
   }
-
-  private val nullFixedValue: FixedExpressionValue = FixedExpressionValue("", "")
 
   private def extractFromAnnotations(param: ParameterData): Option[ParameterEditor] = {
     val dualEditorAnnotation: Option[DualEditor]     = param.getAnnotation[DualEditor]
