+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/definition/component/parameter/editor/ParameterTypeEditorDeterminer.scala
@@ -1,10 +1,10 @@
 package pl.touk.nussknacker.engine.definition.component.parameter.editor
 
 import java.time.temporal.ChronoUnit
-
 import pl.touk.nussknacker.engine.api.definition._
 import pl.touk.nussknacker.engine.api.editor.DualEditorMode
 import pl.touk.nussknacker.engine.api.typed.typing.{SingleTypingResult, TypingResult}
+import pl.touk.nussknacker.engine.graph.expression.FixedExpressionValue
 
 class ParameterTypeEditorDeterminer(val typ: TypingResult) extends ParameterEditorDeterminer {
