+++ interpreter/src/test/scala/pl/touk/nussknacker/engine/definition/component/parameter/editor/EditorExtractorTest.scala
@@ -12,6 +12,7 @@ import pl.touk.nussknacker.engine.api.component.ParameterConfig
 import pl.touk.nussknacker.engine.api.typed.typing.Typed
 import pl.touk.nussknacker.engine.definition.clazz.JavaSampleEnum
 import pl.touk.nussknacker.engine.definition.component.parameter.ParameterData
+import pl.touk.nussknacker.engine.graph.expression.FixedExpressionValue
 
 import scala.reflect.ClassTag
