+++ interpreter/src/test/scala/pl/touk/nussknacker/engine/definition/model/ModelDefinitionFromConfigCreatorExtractorSpec.scala
@@ -10,7 +10,6 @@ import pl.touk.nussknacker.engine.api.component.{ComponentInfo, ComponentType}
 import pl.touk.nussknacker.engine.api.context.{ContextTransformation, ValidationContext}
 import pl.touk.nussknacker.engine.api.definition.{
   AdditionalVariableProvidedInRuntime,
-  FixedExpressionValue,
   FixedValuesValidator,
   MandatoryParameterValidator,
   Parameter,
@@ -26,7 +25,7 @@ import pl.touk.nussknacker.engine.definition.component.methodbased.{
   MethodBasedComponentDefinitionWithImplementation,
   MethodBasedComponentImplementationInvoker
 }
-import pl.touk.nussknacker.engine.graph.expression.Expression
+import pl.touk.nussknacker.engine.graph.expression.{Expression, FixedExpressionValue}
 import pl.touk.nussknacker.engine.modelconfig.{
   ComponentsUiConfigParser,
   DefaultModelConfigLoader,
