+++ scenario-api/src/test/scala/pl/touk/nussknacker/engine/graph/FragmentParameterSerializationSpec.scala
@@ -3,17 +3,9 @@ package pl.touk.nussknacker.engine.graph
 import io.circe.jawn.decode
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
-import pl.touk.nussknacker.engine.graph.expression.Expression
-import pl.touk.nussknacker.engine.graph.node.{
-  FragmentInputDefinition,
-  ParameterValueCompileTimeValidation,
-  ValueInputWithFixedValuesProvided
-}
-import pl.touk.nussknacker.engine.graph.node.FragmentInputDefinition.{
-  FixedExpressionValue,
-  FragmentClazzRef,
-  FragmentParameter
-}
+import pl.touk.nussknacker.engine.graph.expression.{Expression, FixedExpressionValue}
+import pl.touk.nussknacker.engine.graph.node.FragmentInputDefinition.{FragmentClazzRef, FragmentParameter}
+import pl.touk.nussknacker.engine.graph.node.{ParameterValueCompileTimeValidation, ValueInputWithFixedValuesProvided}
 
 class FragmentParameterSerializationSpec extends AnyFunSuite with Matchers {
 
@@ -93,8 +85,8 @@ class FragmentParameterSerializationSpec extends AnyFunSuite with Matchers {
         valueEditor = Some(
           ValueInputWithFixedValuesProvided(
             fixedValuesList = List(
-              FragmentInputDefinition.FixedExpressionValue("'someValue'", "someValue"),
-              FragmentInputDefinition.FixedExpressionValue("'someOtherValue'", "someOtherValue")
+              FixedExpressionValue("'someValue'", "someValue"),
+              FixedExpressionValue("'someOtherValue'", "someOtherValue")
             ),
             allowOtherValue = true
           )
