+++ utils/schemed-kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/schemedkafka/KafkaUniversalComponentTransformer.scala
@@ -8,12 +8,14 @@ import pl.touk.nussknacker.engine.api.context.transformation.{
   DefinedEagerParameter,
   SingleInputGenericNodeTransformation
 }
-import pl.touk.nussknacker.engine.api.definition.{FixedExpressionValue, FixedValuesParameterEditor, Parameter}
+import pl.touk.nussknacker.engine.api.definition.{FixedValuesParameterEditor, Parameter}
 import pl.touk.nussknacker.engine.api.process.ProcessObjectDependencies
-import pl.touk.nussknacker.engine.schemedkafka.KafkaUniversalComponentTransformer.{TopicParamName, nullTopicOption}
+import pl.touk.nussknacker.engine.schemedkafka.KafkaUniversalComponentTransformer.TopicParamName
 import pl.touk.nussknacker.engine.schemedkafka.schemaregistry._
 import pl.touk.nussknacker.engine.api.NodeId
 import pl.touk.nussknacker.engine.api.validation.ValidationMode
+import pl.touk.nussknacker.engine.graph.expression.FixedExpressionValue
+import pl.touk.nussknacker.engine.graph.expression.FixedExpressionValue.nullFixedValue
 import pl.touk.nussknacker.engine.kafka.validator.WithCachedTopicsExistenceValidator
 import pl.touk.nussknacker.engine.kafka.{KafkaComponentsUtils, KafkaConfig, PreparedKafkaTopic}
 import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.universal.UniversalSchemaSupportDispatcher
@@ -28,10 +30,6 @@ object KafkaUniversalComponentTransformer {
 
   def extractValidationMode(value: String): ValidationMode =
     ValidationMode.fromString(value, SinkValidationModeParameterName)
-
-  // Initially we don't want to select concrete topic by user so we add null topic on the beginning of select box.
-  // TODO: add addNullOption feature flag to FixedValuesParameterEditor
-  private val nullTopicOption: FixedExpressionValue = FixedExpressionValue("", "")
 
 }
 
@@ -80,7 +78,9 @@ trait KafkaUniversalComponentTransformer[T]
     Parameter[String](topicParamName).copy(editor =
       Some(
         FixedValuesParameterEditor(
-          nullTopicOption +: topics
+          // Initially we don't want to select concrete topic by user so we add null topic on the beginning of select box.
+          // TODO: add addNullOption feature flag to FixedValuesParameterEditor
+          nullFixedValue +: topics
             .flatMap(topic =>
               processObjectDependencies.objectNaming
                 .decodeName(topic, processObjectDependencies.config, KafkaComponentsUtils.KafkaTopicUsageKey)
